Split mechanics implementation
