+++ lib/spells/CreatureSpellMechanics.h
@@ -11,3 +11,29 @@
  #pragma once
  
  #include "CDefaultSpellMechanics.h"
+
+class AcidBreathDamageMechanics: public DefaultSpellMechanics
+{
+public:
+	AcidBreathDamageMechanics(CSpell * s): DefaultSpellMechanics(s){};
+protected:
+	void applyBattleEffects(const SpellCastEnvironment * env, BattleSpellCastParameters & parameters, SpellCastContext & ctx) const override;		
+};
+
+class DeathStareMechanics: public DefaultSpellMechanics
+{
+public:
+	DeathStareMechanics(CSpell * s): DefaultSpellMechanics(s){};
+protected:
+	void applyBattleEffects(const SpellCastEnvironment * env, BattleSpellCastParameters & parameters, SpellCastContext & ctx) const override;		
+};
+
+class DispellHelpfulMechanics: public DefaultSpellMechanics
+{
+public:
+	DispellHelpfulMechanics(CSpell * s): DefaultSpellMechanics(s){};
+	
+	void applyBattle(BattleInfo * battle, const BattleSpellCast * packet) const override;
+	
+	ESpellCastProblem::ESpellCastProblem isImmuneByStack(const CGHeroInstance * caster, const CStack * obj) const override;	
+};
